+++ app/service-registry.js
@@ -17,8 +17,8 @@ constructor()
  *
  * @param {string} id exchange id
  * @param {string} name exchange name
- * @param {string} instance Exchange object
- * @param {array} features dictionary of features {string:{enabled:boolean}} (optional)
+ * @param {object} instance Exchange object
+ * @param {object} features dictionary of features {string:{enabled:boolean}} (optional)
  * @param {boolean} demoMode indicates whether or not demo mode is enabled for this exchange
  */
 registerExchange(id, name, instance, features, demoMode, obj)
@@ -41,6 +41,11 @@ registerExchange(id, name, instance, features, demoMode, obj)
     this._services.exchanges[id] = {id:id,instance:instance,name:name,features:featureList,demo:demo,dummy:instance.isDummy()}
 }
 
+getExchanges(id)
+{
+    return this._services.exchanges;
+}
+
 getExchange(id)
 {
     if (undefined === this._services.exchanges[id])
@@ -55,10 +60,11 @@ getExchange(id)
  *
  * @param {string} id service id
  * @param {string} name service name
- * @param {array} features dictionary of features {string:{enabled:boolean}} (optional)
+ * @param {object} instance service object
+ * @param {object} features dictionary of features {string:{enabled:boolean}} (optional)
  * @param {boolean} demoMode indicates whether or not demo mode is enabled for this service
  */
-registerService(id, name, features, demoMode)
+registerService(id, name, instance, features, demoMode)
 {
     let featureList = {};
     let demo = false;
@@ -75,12 +81,21 @@ registerService(id, name, features, demoMode)
             demo = true;
         }
     }
-    this._services.others[id] = {id:id,name:name,features:featureList,demo:demo}
+    this._services.others[id] = {id:id,name:name,instance:instance,features:featureList,demo:demo}
 }
 
 getServices()
 {
     return this._services;
+}
+
+getService(id)
+{
+    if (undefined === this._services.others[id])
+    {
+        return null;
+    }
+    return this._services.others[id];
 }
 
 }